Using gulp-jasmine for testing
As it does not fail on error, so you can keep it running in a gulp watch command. 
Also added a vscode launcher task for it, so I can debug the test from within vscode.
NOTE: Breakpoints need to be set in the generated js files, although source mapping is true.